Iris Flower Classification using Support Vector Machines
The Iris flower dataset is a well-known and widely used dataset in the field of machine learning, specifically for classification tasks. The dataset consists of 150 samples of Iris flowers, each belonging to one of three species: Iris-setosa, Iris-versicolor, or Iris-virginica. The goal of this project is to develop a machine learning model capable of accurately classifying Iris flowers based on their sepal length, sepal width, petal length, and petal width. We will employ the Support Vector Machines (SVM) algorithm for this classification task and compare the performance of two different SVM models.