Tech Industry Dynamics: An Investigation into Job Openings and Programming Language Demand Part Two: The Popularity Analysis & Statistical Hypothesis
In the second part of this investigative journey into the heart of the tech industry, we dive into a comprehensive popularity analysis of various programming languages. The ebb and flow of programming language popularity is often indicative of broader tech industry trends, which can, in turn, offer valuable insights into the dynamics of job market demand.